Move more tests to integration
