+++ tests/python/integration/test_integrated_utilities_jurisdictions.py
@@ -1,0 +1,298 @@
+"""COMPASS Ordinance jurisdiction utilities tests"""
+
+from pathlib import Path
+
+import pytest
+import numpy as np
+import pandas as pd
+
+from compass.utilities.jurisdictions import (
+    load_all_jurisdiction_info,
+    load_jurisdictions_from_fp,
+    jurisdiction_websites,
+)
+from compass.exceptions import COMPASSValueError
+from compass.warn import COMPASSWarning
+
+
+def test_load_all_jurisdictions():
+    """Test the `load_all_jurisdiction_info` function"""
+
+    jurisdiction_info = load_all_jurisdiction_info()
+    assert not jurisdiction_info.empty
+
+    expected_cols = [
+        "County",
+        "State",
+        "Subdivision",
+        "Jurisdiction Type",
+        "FIPS",
+        "Website",
+    ]
+    assert all(col in jurisdiction_info for col in expected_cols)
+    for g, data in jurisdiction_info.groupby(
+        ["County", "State", "Subdivision", "Jurisdiction Type"]
+    ):
+        if len(data) > 1:
+            print(g)
+            print(data)
+    assert len(jurisdiction_info) == len(
+        jurisdiction_info.groupby(
+            ["County", "State", "Subdivision", "Jurisdiction Type"]
+        )
+    )
+    assert len(jurisdiction_info) == len(jurisdiction_info.groupby(["FIPS"]))
+
+    # Spot checks:
+    assert "Decatur" in set(jurisdiction_info["County"])
+    assert "Box Elder" in set(jurisdiction_info["County"])
+    assert "Colorado" in set(jurisdiction_info["State"])
+    assert "Rhode Island" in set(jurisdiction_info["State"])
+
+
+def test_jurisdiction_websites():
+    """Test the `jurisdiction_websites` function"""
+
+    websites = jurisdiction_websites()
+    assert len(websites) == len(load_all_jurisdiction_info())
+    assert isinstance(websites, dict)
+
+    # Spot checks:
+    assert 18031 in websites  # Decatur Indiana
+    assert 8041 in websites  # El Paso, Colorado
+    assert 49003 in websites  # Box Elder, Utah
+
+
+def test_load_jurisdictions_from_fp(tmp_path):
+    """Test `load_jurisdictions_from_fp` function"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {"County": ["decatur", "DNE County"], "State": ["INDIANA", "colorado"]}
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    with pytest.warns(COMPASSWarning) as record:
+        jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(record) == 1
+    warning_msg = str(record[0].message)
+
+    assert "nan" not in warning_msg
+    assert "DNE County" in warning_msg
+    assert "colorado" in warning_msg
+
+    assert len(jurisdictions) == 1
+    assert set(jurisdictions["County"]) == {"Decatur"}
+    assert set(jurisdictions["State"]) == {"Indiana"}
+    assert set(jurisdictions["Subdivision"]) == {None}
+    assert set(jurisdictions["Subdivision"]) != {np.nan}
+    assert set(jurisdictions["Jurisdiction Type"]) == {"county"}
+    assert {type(val) for val in jurisdictions["FIPS"]} == {str}
+
+
+def test_load_jurisdictions_from_fp_bad_input(tmp_path):
+    """Test `load_jurisdictions_from_fp` function"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    pd.DataFrame().to_csv(test_jurisdiction_fp)
+
+    with pytest.raises(COMPASSValueError) as err:
+        load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    expected_msg = (
+        "The jurisdiction input must have at least a 'State' column!"
+    )
+    assert expected_msg in str(err)
+
+
+def test_load_jurisdictions_from_fp_single_county(tmp_path):
+    """Test that `load_jurisdictions_from_fp` returns a single county"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {"County": ["Wharton"], "State": ["Texas"]}
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 1
+    assert set(jurisdictions["County"]) == {"Wharton"}
+    assert set(jurisdictions["State"]) == {"Texas"}
+    assert set(jurisdictions["Subdivision"]) == {None}
+    assert set(jurisdictions["Subdivision"]) != {np.nan}
+    assert set(jurisdictions["Jurisdiction Type"]) == {"county"}
+    assert {type(val) for val in jurisdictions["FIPS"]} == {str}
+
+
+def test_load_jurisdictions_no_repeated_counties(tmp_path):
+    """Test that `load_jurisdictions_from_fp` doesn't have repeats"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": ["Jefferson", "Jefferson", "Jefferson"],
+            "State": ["Alabama", "Colorado", "Alabama"],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 2
+    assert set(jurisdictions["County"]) == {"Jefferson"}
+    assert set(jurisdictions["State"]) == {"Alabama", "Colorado"}
+    assert set(jurisdictions["Subdivision"]) == {None}
+    assert set(jurisdictions["Subdivision"]) != {np.nan}
+    assert set(jurisdictions["Jurisdiction Type"]) == {"county"}
+    assert {type(val) for val in jurisdictions["FIPS"]} == {str}
+
+
+def test_load_jurisdictions_no_repeated_townships(tmp_path):
+    """Test that `load_jurisdictions_from_fp` doesn't have repeats"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": "Aroostook",
+            "State": "Maine",
+            "Subdivision": ["Perham", "Oakfield", "Perham"],
+            "Jurisdiction Type": "town",
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 2
+    assert set(jurisdictions["County"]) == {"Aroostook"}
+    assert set(jurisdictions["State"]) == {"Maine"}
+    assert set(jurisdictions["Subdivision"]) == {"Perham", "Oakfield"}
+    assert set(jurisdictions["Jurisdiction Type"]) == {"town"}
+    assert {type(val) for val in jurisdictions["FIPS"]} == {str}
+
+
+def test_load_jurisdictions_no_repeated_townships_and_counties(tmp_path):
+    """Test that `load_jurisdictions_from_fp` doesn't have repeats"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": "Aroostook",
+            "State": "Maine",
+            "Subdivision": ["Perham", "Oakfield", "Perham", None, None],
+            "Jurisdiction Type": ["town", "town", "town", "county", "county"],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 3
+    assert set(jurisdictions["County"]) == {"Aroostook"}
+    assert set(jurisdictions["State"]) == {"Maine"}
+    assert set(jurisdictions["Subdivision"]) == {"Perham", "Oakfield", None}
+    assert set(jurisdictions["Jurisdiction Type"]) == {"town", "county"}
+    assert {type(val) for val in jurisdictions["FIPS"]} == {str}
+
+
+def test_load_jurisdictions_from_fp_missing_jurisdiction_type(tmp_path):
+    """Test error when Subdivision provided without Jurisdiction Type column"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": ["Aroostook"],
+            "State": ["Maine"],
+            "Subdivision": ["Perham"],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    with pytest.raises(COMPASSValueError) as exc_info:
+        load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    error_msg = str(exc_info.value)
+    assert "Jurisdiction Type" in error_msg
+    assert "Subdivision" in error_msg
+    assert "must have" in error_msg
+
+
+def test_load_jurisdictions_from_fp_warning_message_content(tmp_path):
+    """Test that warning message contains key information"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": ["Fake County", "Another Fake"],
+            "State": ["Colorado", "Texas"],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    with pytest.warns(COMPASSWarning) as record:
+        jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(record) == 1
+    warning_msg = str(record[0].message)
+
+    assert "not found" in warning_msg.lower()
+    assert "Fake County" in warning_msg
+    assert "Another Fake" in warning_msg
+    assert "Colorado" in warning_msg
+    assert "Texas" in warning_msg
+    assert (
+        "spelling" in warning_msg.lower()
+        or "capitalization" in warning_msg.lower()
+    )
+
+    assert len(jurisdictions) == 0
+
+
+def test_load_jurisdictions_from_fp_whitespace_trimming(tmp_path):
+    """Test that input jurisdictions have whitespace trimmed"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": ["  Decatur  ", "Wharton"],
+            "State": ["  Indiana  ", "  Texas  "],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 2
+    assert "Decatur" in set(jurisdictions["County"])
+    assert "Wharton" in set(jurisdictions["County"])
+    assert "Indiana" in set(jurisdictions["State"])
+    assert "Texas" in set(jurisdictions["State"])
+
+
+def test_load_jurisdictions_from_fp_subdivision_whitespace_trimming(tmp_path):
+    """Test whitespace trimming for subdivisions and jurisdiction types"""
+
+    test_jurisdiction_fp = tmp_path / "out.csv"
+    input_jurisdictions = pd.DataFrame(
+        {
+            "County": ["  Aroostook  "],
+            "State": ["  Maine  "],
+            "Subdivision": ["  Perham  "],
+            "Jurisdiction Type": ["  town  "],
+        }
+    )
+    input_jurisdictions.to_csv(test_jurisdiction_fp)
+
+    jurisdictions = load_jurisdictions_from_fp(test_jurisdiction_fp)
+
+    assert len(jurisdictions) == 1
+    assert jurisdictions.iloc[0]["County"] == "Aroostook"
+    assert jurisdictions.iloc[0]["State"] == "Maine"
+    assert jurisdictions.iloc[0]["Subdivision"] == "Perham"
+    assert jurisdictions.iloc[0]["Jurisdiction Type"] == "town"
+
+
+if __name__ == "__main__":
+    pytest.main(["-q", "--show-capture=all", Path(__file__), "-rapP"])
